Adds better handling of model description
The standard adopted was clashing with brian2 implementation, so I created a preprocessing stage that removes the problematic part. This should be fine because the original motivation for it was that customizing the models sometimes leads to unnoticed/undesired consequences. However, building the  model is well handled by brian2. One example of this was a synaptic pathway being added to model descript ('pre': 'spike') due to how a previous NeuronGroup that was built. Moreover, simply creating a new dict entry is also done, which is not possible with ParamDict.